tgusurvey/tgusurvey/src/test/javascript/spec/app/entities/surveystore/education-program/education-program.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_FORMAT } from 'app/shared/constants/input.constants';
import { EducationProgramService } from 'app/entities/surveystore/education-program/education-program.service';
import { IEducationProgram, EducationProgram } from 'app/shared/model/surveystore/education-program.model';

describe('Service Tests', () => {
  describe('EducationProgram Service', () => {
    let injector: TestBed;
    let service: EducationProgramService;
    let httpMock: HttpTestingController;
    let elemDefault: IEducationProgram;
    let expectedResult: IEducationProgram | IEducationProgram[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(EducationProgramService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new EducationProgram(0, 'AAAAAAA', 'AAAAAAA', currentDate);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            year: currentDate.format(DATE_FORMAT),
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a EducationProgram', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            year: currentDate.format(DATE_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            year: currentDate,
          },
          returnedFromService
        );

        service.create(new EducationProgram()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a EducationProgram', () => {
        const returnedFromService = Object.assign(
          {
            msct: 'BBBBBB',
            name: 'BBBBBB',
            year: currentDate.format(DATE_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            year: currentDate,
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of EducationProgram', () => {
        const returnedFromService = Object.assign(
          {
            msct: 'BBBBBB',
            name: 'BBBBBB',
            year: currentDate.format(DATE_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            year: currentDate,
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a EducationProgram', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
